requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
numpy
opencv-python-headless
google-genai
paddlepaddle
paddleocr
